feedback from v bdhrs

+++ build.ps1
@@ -1,4 +1,11 @@
-python .\scripts\xlsx2csv.py .\_languages.xlsx ./build --all
+$pyCmd = Get-Command "python.exe" -ErrorAction SilentlyContinue
+if (-not ($pyCmd.Version -gt "3.0"))
+{
+  Write-Host -ForegroundColor Red "Unable to find python v3. Quitting..."
+  return
+}
+
+python.exe .\scripts\xlsx2csv.py .\_languages.xlsx ./build --all
 
 Remove-Item "./build/* - notes.csv"
